max/min freq
Continued work on setting the min/max frequencies of the display. The current method is to calculate the fft of the entire waveform and find the line and scatter that are within min/max. The problem with this is that the pan and zoom do not work correctly.
Another approach not limit the data for display purposes but only limit the data for reporting of the peaks. This allows the pan/and zoom to work but requires us to control the default setting of the pan and zoom as well as allow a zoom out/zoom in function. Hummm