Rework
The setting of min/max now only effects the range used for magnitude thresholding and reporting of peaks. It has no effect on zoom but it is used to set the "home" zoom view. pan and zoom can now see the complete fft sequence and any peaks that have been marked on that sequence.