Rework meaning of hold, updates to threshold_slider, and fix home
Changed Peak Hold to Hold Result so that averaging is only enabled when not holding result.
Added override on home from the NavigationToolbar to reset the Y axis to -100 to 0.
Added sunken border and major tick marks. Made all tick marks go to the end of the slider.
Update the Peaks when changing threshold or restting the x_limits.
Return mag and mag_db from the dft analysis instead of mag_db and complex. Allows removal of code to calculate the mag from complex.