Mac installer
Worked through dmg, zip, and pkg. Result is to use pkg. Added commands to make the package, notarize the package, and staple the package. The developer id's used in the mac_ and the guitar-tap.spec need to be changed for different users. Removed the makedmg script since it failed to notarize.