Started work on annotations
Refactored code to make things cleaner.
Added column to table for the modes and use a combobox to select the modes.
Redid the hold results logic so that the data model in the table and the state of the table is cleared when NOT holding the results. All selection and editing is also disabled when NOT holding results.
A fair amount of clean-up and documentation is required.
Also added Help for the toolbar and disabled confusing icons on the toolbar.
Reordered the setting of the number of averages before enabling it - always seemed to get it backwards when using it.