Work on selection retention across hold results.
Changed saved peak as frequency instead of index so that we can handle hold results transitions if there is no new data.
Use datachanged signal from setting mode or show to make sure the row and peak are selected.
Major goal was to allow for retention of selected frequency when hold results disabled and then enabled.