The text of the data model was bleeding through the editor cells for mode and show.
The model.data routine was returning the value of the editor. Since these values are actually saved in a different place (i.e. through show_value/mode_value) the data function can return blanks. Had to modify the setEditorData to get the value from data_value instead of data. Which required adding data_value to the filter_model along with index translation from the proxy to the source. The bleeding text is now gone.